# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel API",
        "description": "Hotel API Documentation",
        "contact": {
            "email": "diogo.coutinho.ads@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/hotels/{hotel}/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get list of rooms",
                "description": "Returns list of rooms",
                "operationId": "ecddddaf367277fcd55659008183d69f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create new room",
                "description": "Create new room and return room data",
                "operationId": "18f78f1dffdeeeaf402fdd5ff2f59352",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRoomFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hotels/{hotel}/rooms/{room}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get room details",
                "description": "Returns room details",
                "operationId": "713332d968e211262ba0a7dff331250f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update room",
                "description": "Update room and return room data",
                "operationId": "7ce92b5a1a352409155df8538141ebdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRoomFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete room",
                "description": "Delete room",
                "operationId": "77cd716b7b5f458448da57569fa1bca0",
                "responses": {
                    "200": {
                        "description": "Room deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Room deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreRoomFormRequest": {
                "title": "StoreRoomFormRequest",
                "description": "Store Room Form Request",
                "required": [
                    "name",
                    "description",
                    "hotel_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Room 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Room 1 description"
                    },
                    "hotel_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "UpdateRoomFormRequest": {
                "title": "UpdateRoomFormRequest",
                "description": "Update Room Form Request",
                "required": [
                    "name",
                    "description",
                    "hotel_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Room 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Room 1 description"
                    },
                    "hotel_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "RoomResource": {
                "title": "RoomResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "hotel_id": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum Token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}